Move getElementPosition to Editor
